Remove dead code and notebooks
