+++ moscot/problems/_anndata.py
@@ -6,8 +6,8 @@ import numpy.typing as npt
 
 from anndata import AnnData
 
-from moscot.utils import _get_backend_losses
-from moscot._costs import __all__ as moscot_losses, BaseLoss
+from moscot._utils import _get_backend_losses
+from moscot.costs._costs import __all__ as moscot_losses, BaseLoss
 from moscot.solvers._tagged_arry import Tag, TaggedArray
