--- old_code_to_recycle/framework/settings.py
@@ -1,4 +1,0 @@
-from typing import Union, Literal
-
-# TODO: check where we want to save stuff like this to prevent circular imports
-strategies_MatchingEstimator = Union[Literal["pairwise"], Literal["sequential"]]
